requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.43.1
blingfire==0.1.8
blis==1.2.0
blobfile==2.1.1
BM25==1.0.0
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
datefinder==0.7.3
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
emoji==2.14.1
evaluate==0.4.2
filelock==3.15.4
fire==0.6.0
flash_attn==2.7.4.post1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hf_transfer==0.1.6
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.25.1
lmdb==1.6.2
loguru==0.7.2
lxml==4.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.12
mwparserfromhell==0.6.6
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.13.3
packaging==24.1
pandas==2.2.2
peft==0.14.0
pillow==10.4.0
platformdirs==4.3.2
portalocker==2.10.1
preshed==3.0.9
protobuf==5.28.0
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge==1.0.1
rouge_score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==69.5.1
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stanza==1.10.1
sympy==1.12.1
tabulate==0.9.0
tagme==0.1.3
termcolor==2.4.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.3.1
tqdm==4.66.4
transformers==4.48.3
trl==0.14.0
typer==0.15.1
typing_extensions==4.12.2
tyro==0.8.8
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.9
wasabi==1.1.3
weasel==0.4.1
wheel==0.43.0
wikipedia2vec==2.0.0
wikitables==0.5.5
wrapt==1.17.2
xxhash==3.4.1
yarl==1.9.4
